refactor: rename kluster1 directory and package to kubecheck
Complete the rebranding by renaming:
- Directory: kluster1/ → kubecheck/
- Files: kluster1.go → kubecheck.go, kluster1_test.go → kubecheck_test.go
- Package names: package kluster1 → package kubecheck
- Import paths: kubecheck/kluster1 → kubecheck/kubecheck

BREAKING CHANGE: Package name changed from kluster1 to kubecheck. Users must update their imports.
