+++ kubecheck/kubecheck.go
@@ -1,11 +1,11 @@
-package kluster1
+package kubecheck
 
 import (
 	"fmt"
 	"log"
 
-	"github.com/ogarciacar/kubecheck/kluster1/compute/runtime/kindcluster"
-	"github.com/ogarciacar/kubecheck/kluster1/storage/persistence/tempkubeconfig"
+	"github.com/ogarciacar/kubecheck/kubecheck/compute/runtime/kindcluster"
+	"github.com/ogarciacar/kubecheck/kubecheck/storage/persistence/tempkubeconfig"
 	"github.com/ogarciacar/kubecheck/sdk"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
