+++ kubecheck/quick_start_test.go
@@ -1,9 +1,9 @@
-package kluster1_test
+package kubecheck_test
 
 import (
 	"testing"
 
-	"github.com/ogarciacar/kubecheck/kluster1"
+	"github.com/ogarciacar/kubecheck/kubecheck"
 )
 
 func TestMyKubernetesApp(t *testing.T) {
